Backend/requirements.txt
aiofiles
click==7.1.2
cycler==0.10.0
uvicorn
python-multipart
pillow
numpy
pandas
scikit-learn
fastapi
pydantic
starlette
efficientnet-pytorch==0.7.1
tqdm